+++ src/app/login/page.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import Image from 'next/image';
-import LoginForm from './LoginForm';
+import LoginForm from './components/LoginForm';
 
 export default function LoginPage() {
   return (
